+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 **Sample App**
 - **Feature:** User can edit contact name and email in App Settings.
+- **Feature:** User can see access token scopes requested for the RAE audience in first-time screen.
 
 ### 3.5.0 (2021-07-27)
 **SDK**